refactor(web): 抽出通用会话快照与墙部件，收掉几处照抄
墙的表现层早就是公共的（海报格子、图廊、滚动恢复、位置记忆），但它们周围
反复照抄的两样东西一直没人收：

1. lib/session-snapshot.ts —— 带上限的会话快照仓库。
   「先 delete 再 set、超限淘汰队首」这段同样的 LRU 全站有三份：滚动位置
   （lib/use-scroll-restoration.ts）、单库页快照（lib/library-detail-snapshot.ts）、
   片单页快照（components/collection-grid-view.tsx）。收成一份
   createSessionSnapshots<K, V>(capacity)，三处改为引用，上限各自不变
   （100 / 20 / 32）。纯逻辑模块，配 node --test 覆盖淘汰顺序。

2. components/wall-chrome.tsx —— 一面墙的三个通用部件。
   WallLoadMore / WallLoadPrev / WallRecallPill 原先长在单库页文件里，收藏页
   接同一套位置记忆时得从那个页面文件 import——公共件寄居在某一个页面里，
   谁用都要绕道。搬出来之后两页平级引用，单库页少了 140 行。

纯搬运与去重，没有行为改动：typecheck / lint（新文件零问题）/ node --test /
next build 全通过。

按 offset 分页的窗口状态机（loadMore / loadPrev / jumpTo / reload 与补页的滚动
补偿）两页仍是各一份，那是下一个 PR 的事，要连同 docs/design/media-wall.md
把契约一起写下来。
